post _remove

+++ blog/views.py
@@ -102,7 +102,6 @@ def logout_request(request):
 
 def post_remove(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    post.delete()
-    return redirect('post_list')
-
-
+    if request.user ==post.author:
+       post.delete()
+       return redirect('post_list')
